Restructure header components
