+++ frontend/src/components/header/_styles.css
@@ -7,106 +7,6 @@
   width: 100%;
 }
 
-.header__video {
-  z-index: -1;
-  position: absolute;
-  width: 100%;
-  max-height: var(--max-header-height);
-  object-fit: cover;
-  min-height: 550px;
-}
-
-.header__content__title {
-  margin: 0;
-  color: white;
-  font-size: 84px;
-  text-transform: uppercase;
-  font-weight: 800;
-
-  @media (max-width: 1250px) {
-    font-size: 71px;
-  }
-
-  @media (max-width: 1040px) {
-    font-size: 61px;
-  }
-
-  &--first {
-    margin: 87px 0 -15px 0;
-
-    @media (max-width: 1400px) {
-      margin-top: 50px;
-    }
-
-    @media (max-width: 1250px) {
-      margin-top: 22px;
-    }
-  }
-}
-
-.header__content__description {
-  color: var(--gallery);
-  max-width: var(--max-width);
-  margin-top: 30px;
-  font-size: 24px;
-  font-weight: 700;
-
-  @media (max-width: 1250px) {
-    font-size: 18px;
-    margin-top: 20px;
-  }
-
- @media (max-width: 865px) {
-    display: none;
-  }
-}
-
-.header__buttons {
-  margin-top: 65px;
-  display: grid;
-  grid-auto-flow: column;
-  grid-gap: 20px;
-  justify-items: left;
-  grid-template-columns: 214px auto;
-
-  @media (max-width: 1400px) {
-    margin-top: 30px;
-  }
-}
-
-.header__buttons__apply {
-  text-transform: uppercase;
-  font-size: 18px;
-  width: 214px;
-  height: 50px;
-  border: 4px solid white;
-  background-color: white;
-  color: black;
-  font-weight: 700;
-  text-align: center;
-  cursor: pointer;
-  transition: all 0.3s;
-
-  &:hover {
-    background-color: black;
-    color: var(--cerulean);
-    border: 4px solid black;
-  }
-}
-
-.header__buttons__more {
-  width: 329px;
-  height: 50px;
-  border: 4px solid white;
-  font-family: Montserrat;
-  background-color: transparent;
-  font-size: 18px;
-  color: white;
-  font-weight: 700;
-  text-align: center;
-  cursor: pointer;
-}
-
 .header__more-icon {
   width: 70px;
   margin-top: 75px;
