+++ frontend/src/styles.css
@@ -3,7 +3,9 @@
 @import url('./_colors.css');
 
 @import url("./components/header/_styles.css");
-@import url("./components/nav/_styles.css");
+@import url("./components/header/content/_styles.css");
+@import url("./components/header/cover-video/_styles.css");
+@import url("./components/header/nav/_styles.css");
 
 
 html, body, #app {
